+++ borrar/Obstaculo.cpp
