+++ borrar/Obstaculo.h
