+++ src/PantallaMenu.cpp
@@ -1,6 +1,9 @@
 #include "../include/PantallaMenu.h"
+#include "../include/PantallaSeleccionar.h"
+#include "../include/Juego.h"
+#include <memory>
 
-PantallaMenu::PantallaMenu(std::function<void()> onStartCallback) : onStart(onStartCallback){
+PantallaMenu::PantallaMenu(Juego* juego) : Pantalla(juego){
     fuente.loadFromFile("../assets/Bangers-Regular.ttf");
     texto.setFont(fuente);
     texto.setString("Presiona Enter para Comenzar!");
@@ -10,7 +13,7 @@ PantallaMenu::PantallaMenu(std::function<void()> onStartCallback) : onStart(onSt
 
 void PantallaMenu::ManejarEvento(sf::Event evento){
     if (evento.type == sf::Event::KeyPressed && evento.key.code == sf::Keyboard::Enter){
-        onStart();
+        juego->cambiarPantalla(std::make_unique<PantallaSeleccionar>(juego));
     }
 }
