--- src/principal.cpp
@@ -1,7 +1,0 @@
-#include "../include/Juego.h"
-
-int main(){
-    Juego Juego;
-    Juego.run();
-    return 0;
-}
